Added an encoder, formatter and fixed comma error
This marks the end of this tiny project, there are alot of improvements
to be done and bugs to be fixed. The trailing comma algorithm has been
fixed. Added a very minimal formatter to format json strings, still
needs a ton of improvement, the interface is poor and very shaky and
also added a simple encoder, shaky as well, I have not added proper
exceptions, this is the next step and also write a bunch of tests, yeah.
But for now, project is complete. It can load json content into python
and dump from python objects to json string. Amaaaaazing.