Type bugs and appreciation
This is probably my last commit in this repo as the project is
officially done. I'm having a hard time understanding python types. I
mean if we have 3 classes A, B and C where B inherits from A and C from
B doesn't that mean that all c's are A's and B's. C will inherit all
properties and methods of A and B hence C is an A and a B. This also
applies to B and A, all B's are A's. But for whatever reason my static
analysis is flagging errors when I pass a C in place of B, do they
expect me to pass an A all should it always be a B. This has to do with
covariance and contravariance or it is just a bug in the initializer.
More tests of this will be done in the code repo. For now, GoodBye.